Surface API, implementation in Vulkan
