+++ blade-graphics/src/vulkan/mod.rs
@@ -6,6 +6,7 @@ mod descriptor;
 mod init;
 mod pipeline;
 mod resource;
+mod surface;
 
 const QUERY_POOL_SIZE: usize = crate::limits::PASS_COUNT + 1;
 
@@ -40,6 +41,7 @@ struct Workarounds {
 struct Device {
     core: ash::Device,
     device_information: crate::DeviceInformation,
+    swapchain: Option<khr::swapchain::Device>,
     debug_utils: ash::ext::debug_utils::Device,
     timeline_semaphore: khr::timeline_semaphore::Device,
     dynamic_rendering: khr::dynamic_rendering::Device,
@@ -65,41 +67,59 @@ struct Queue {
     last_progress: u64,
 }
 
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct Frame {
-    image_index: u32,
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
+struct InternalFrame {
+    acquire_semaphore: vk::Semaphore,
     image: vk::Image,
     view: vk::ImageView,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+struct Swapchain {
+    raw: vk::SwapchainKHR,
     format: crate::TextureFormat,
-    acquire_semaphore: vk::Semaphore,
     target_size: [u16; 2],
+}
+
+pub struct Surface {
+    raw: vk::SurfaceKHR,
+    frames: Vec<InternalFrame>,
+    next_semaphore: vk::Semaphore,
+    swapchain: Swapchain,
+    _full_screen_exclusive: bool,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+struct Presentation {
+    swapchain: vk::SwapchainKHR,
+    image_index: u32,
+    acquire_semaphore: vk::Semaphore,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct Frame {
+    swapchain: Swapchain,
+    image_index: u32,
+    internal: InternalFrame,
 }
 
 impl Frame {
     pub fn texture(&self) -> Texture {
         Texture {
-            raw: self.image,
+            raw: self.internal.image,
             memory_handle: !0,
-            target_size: self.target_size,
-            format: self.format,
+            target_size: self.swapchain.target_size,
+            format: self.swapchain.format,
         }
     }
 
     pub fn texture_view(&self) -> TextureView {
         TextureView {
-            raw: self.view,
-            target_size: self.target_size,
+            raw: self.internal.view,
+            target_size: self.swapchain.target_size,
             aspects: crate::TexelAspects::COLOR,
         }
     }
-}
-
-struct Surface {
-    raw: vk::SurfaceKHR,
-    frames: Vec<Frame>,
-    next_semaphore: vk::Semaphore,
-    swapchain: vk::SwapchainKHR,
-    extension: khr::swapchain::Device,
 }
 
 fn map_timeout(millis: u32) -> u64 {
@@ -115,12 +135,11 @@ pub struct Context {
     device: Device,
     queue_family_index: u32,
     queue: Mutex<Queue>,
-    surface: Option<Mutex<Surface>>,
     physical_device: vk::PhysicalDevice,
     naga_flags: naga::back::spv::WriterFlags,
     shader_debug_path: Option<PathBuf>,
     instance: Instance,
-    _entry: ash::Entry,
+    entry: ash::Entry,
 }
 
 #[derive(Clone, Copy, Debug, Hash, PartialEq)]
@@ -230,12 +249,6 @@ struct CommandBuffer {
     descriptor_pool: descriptor::DescriptorPool,
     query_pool: vk::QueryPool,
     timed_pass_names: Vec<String>,
-}
-
-#[derive(Debug, PartialEq)]
-struct Presentation {
-    image_index: u32,
-    acquire_semaphore: vk::Semaphore,
 }
 
 struct CrashHandler {
@@ -465,15 +478,15 @@ impl crate::traits::CommandDevice for Context {
         encoder.check_gpu_crash(ret);
 
         if let Some(presentation) = encoder.present.take() {
-            let surface = self.surface.as_ref().unwrap().lock().unwrap();
-            let swapchains = [surface.swapchain];
+            let khr_swapchain = self.device.swapchain.as_ref().unwrap();
+            let swapchains = [presentation.swapchain];
             let image_indices = [presentation.image_index];
             let wait_semaphores = [queue.present_semaphore];
             let present_info = vk::PresentInfoKHR::default()
                 .swapchains(&swapchains)
                 .image_indices(&image_indices)
                 .wait_semaphores(&wait_semaphores);
-            let ret = unsafe { surface.extension.queue_present(queue.raw, &present_info) };
+            let ret = unsafe { khr_swapchain.queue_present(queue.raw, &present_info) };
             let _ = encoder.check_gpu_crash(ret);
         }
