+++ blade-graphics/src/vulkan/surface.rs
@@ -1,0 +1,375 @@
+use ash::vk;
+use std::mem;
+
+impl super::Context {
+    pub fn create_surface<
+        I: raw_window_handle::HasWindowHandle + raw_window_handle::HasDisplayHandle,
+    >(
+        &self,
+        window: &I,
+    ) -> Result<super::Surface, crate::NotSupportedError> {
+        let raw = unsafe {
+            ash_window::create_surface(
+                &self.entry,
+                &self.instance.core,
+                window.display_handle().unwrap().as_raw(),
+                window.window_handle().unwrap().as_raw(),
+                None,
+            )
+            .map_err(|e| crate::NotSupportedError::VulkanError(e))?
+        };
+
+        let khr_surface = self
+            .instance
+            .surface
+            .as_ref()
+            .ok_or(crate::NotSupportedError::PlatformNotSupported)?;
+        if unsafe {
+            khr_surface.get_physical_device_surface_support(
+                self.physical_device,
+                self.queue_family_index,
+                raw,
+            ) != Ok(true)
+        } {
+            log::warn!("Rejected for not presenting to the window surface");
+            return Err(crate::NotSupportedError::PlatformNotSupported);
+        }
+
+        let surface_info = vk::PhysicalDeviceSurfaceInfo2KHR {
+            surface: raw,
+            ..Default::default()
+        };
+        let mut fullscreen_exclusive_ext = vk::SurfaceCapabilitiesFullScreenExclusiveEXT::default();
+        let mut capabilities2_khr =
+            vk::SurfaceCapabilities2KHR::default().push_next(&mut fullscreen_exclusive_ext);
+        let _ = unsafe {
+            self.instance
+                .get_surface_capabilities2
+                .get_physical_device_surface_capabilities2(
+                    self.physical_device,
+                    &surface_info,
+                    &mut capabilities2_khr,
+                )
+        };
+        log::debug!("{:?}", capabilities2_khr.surface_capabilities);
+
+        let semaphore_create_info = vk::SemaphoreCreateInfo::default();
+        let next_semaphore = unsafe {
+            self.device
+                .core
+                .create_semaphore(&semaphore_create_info, None)
+                .unwrap()
+        };
+        Ok(super::Surface {
+            raw,
+            frames: Vec::new(),
+            next_semaphore,
+            swapchain: super::Swapchain {
+                raw: vk::SwapchainKHR::null(),
+                target_size: [0; 2],
+                format: crate::TextureFormat::Rgba8Unorm,
+            },
+            _full_screen_exclusive: fullscreen_exclusive_ext.full_screen_exclusive_supported != 0,
+        })
+    }
+
+    pub fn destroy_surface(&self, surface: &mut super::Surface) {
+        unsafe {
+            self.deinit_swapchain(surface);
+            self.device
+                .core
+                .destroy_semaphore(surface.next_semaphore, None)
+        };
+        if let Some(ref surface_instance) = self.instance.surface {
+            unsafe { surface_instance.destroy_surface(surface.raw, None) };
+        }
+    }
+
+    unsafe fn deinit_swapchain(&self, surface: &mut super::Surface) {
+        let khr_swapchain = self.device.swapchain.as_ref().unwrap();
+        khr_swapchain.destroy_swapchain(mem::take(&mut surface.swapchain.raw), None);
+        for frame in surface.frames.drain(..) {
+            self.device.core.destroy_image_view(frame.view, None);
+            self.device
+                .core
+                .destroy_semaphore(frame.acquire_semaphore, None);
+        }
+    }
+
+    pub fn configure_surface(
+        &self,
+        surface: &mut super::Surface,
+        config: crate::SurfaceConfig,
+    ) -> crate::SurfaceInfo {
+        let khr_swapchain = self.device.swapchain.as_ref().unwrap();
+        let khr_surface = self.instance.surface.as_ref().unwrap();
+
+        let capabilities = unsafe {
+            khr_surface
+                .get_physical_device_surface_capabilities(self.physical_device, surface.raw)
+                .unwrap()
+        };
+        if config.size.width < capabilities.min_image_extent.width
+            || config.size.width > capabilities.max_image_extent.width
+            || config.size.height < capabilities.min_image_extent.height
+            || config.size.height > capabilities.max_image_extent.height
+        {
+            log::warn!(
+                "Requested size {}x{} is outside of surface capabilities",
+                config.size.width,
+                config.size.height
+            );
+        }
+
+        let (alpha, composite_alpha) = if config.transparent {
+            if capabilities
+                .supported_composite_alpha
+                .contains(vk::CompositeAlphaFlagsKHR::POST_MULTIPLIED)
+            {
+                (
+                    crate::AlphaMode::PostMultiplied,
+                    vk::CompositeAlphaFlagsKHR::POST_MULTIPLIED,
+                )
+            } else if capabilities
+                .supported_composite_alpha
+                .contains(vk::CompositeAlphaFlagsKHR::PRE_MULTIPLIED)
+            {
+                (
+                    crate::AlphaMode::PreMultiplied,
+                    vk::CompositeAlphaFlagsKHR::PRE_MULTIPLIED,
+                )
+            } else {
+                log::error!(
+                    "No composite alpha flag for transparency: {:?}",
+                    capabilities.supported_composite_alpha
+                );
+                (
+                    crate::AlphaMode::Ignored,
+                    vk::CompositeAlphaFlagsKHR::OPAQUE,
+                )
+            }
+        } else {
+            (
+                crate::AlphaMode::Ignored,
+                vk::CompositeAlphaFlagsKHR::OPAQUE,
+            )
+        };
+
+        let (requested_frame_count, mode_preferences) = match config.display_sync {
+            crate::DisplaySync::Block => (3, [vk::PresentModeKHR::FIFO].as_slice()),
+            crate::DisplaySync::Recent => (
+                3,
+                [
+                    vk::PresentModeKHR::MAILBOX,
+                    vk::PresentModeKHR::FIFO_RELAXED,
+                    vk::PresentModeKHR::IMMEDIATE,
+                ]
+                .as_slice(),
+            ),
+            crate::DisplaySync::Tear => (2, [vk::PresentModeKHR::IMMEDIATE].as_slice()),
+        };
+        let effective_frame_count = requested_frame_count.max(capabilities.min_image_count);
+
+        let present_modes = unsafe {
+            khr_surface
+                .get_physical_device_surface_present_modes(self.physical_device, surface.raw)
+                .unwrap()
+        };
+        let present_mode = *mode_preferences
+            .iter()
+            .find(|mode| present_modes.contains(mode))
+            .unwrap();
+        log::info!("Using surface present mode {:?}", present_mode);
+
+        let queue_families = [self.queue_family_index];
+
+        let mut supported_formats = Vec::new();
+        let (format, surface_format) = if surface.swapchain.target_size[0] > 0 {
+            let format = surface.swapchain.format;
+            log::info!("Retaining current format: {:?}", format);
+            let vk_color_space = match (format, config.color_space) {
+                (crate::TextureFormat::Bgra8Unorm, crate::ColorSpace::Srgb) => {
+                    vk::ColorSpaceKHR::SRGB_NONLINEAR
+                }
+                (crate::TextureFormat::Bgra8Unorm, crate::ColorSpace::Linear) => {
+                    vk::ColorSpaceKHR::EXTENDED_SRGB_LINEAR_EXT
+                }
+                (crate::TextureFormat::Bgra8UnormSrgb, crate::ColorSpace::Linear) => {
+                    vk::ColorSpaceKHR::default()
+                }
+                _ => panic!(
+                    "Unexpected format {:?} under color space {:?}",
+                    format, config.color_space
+                ),
+            };
+            (
+                format,
+                vk::SurfaceFormatKHR {
+                    format: super::map_texture_format(format),
+                    color_space: vk_color_space,
+                },
+            )
+        } else {
+            supported_formats = unsafe {
+                khr_surface
+                    .get_physical_device_surface_formats(self.physical_device, surface.raw)
+                    .unwrap()
+            };
+            match config.color_space {
+                crate::ColorSpace::Linear => {
+                    let surface_format = vk::SurfaceFormatKHR {
+                        format: vk::Format::B8G8R8A8_UNORM,
+                        color_space: vk::ColorSpaceKHR::EXTENDED_SRGB_LINEAR_EXT,
+                    };
+                    if supported_formats.contains(&surface_format) {
+                        log::info!("Using linear SRGB color space");
+                        (crate::TextureFormat::Bgra8Unorm, surface_format)
+                    } else {
+                        (
+                            crate::TextureFormat::Bgra8UnormSrgb,
+                            vk::SurfaceFormatKHR {
+                                format: vk::Format::B8G8R8A8_SRGB,
+                                color_space: vk::ColorSpaceKHR::default(),
+                            },
+                        )
+                    }
+                }
+                crate::ColorSpace::Srgb => (
+                    crate::TextureFormat::Bgra8Unorm,
+                    vk::SurfaceFormatKHR {
+                        format: vk::Format::B8G8R8A8_UNORM,
+                        color_space: vk::ColorSpaceKHR::SRGB_NONLINEAR,
+                    },
+                ),
+            }
+        };
+        if !supported_formats.is_empty() && !supported_formats.contains(&surface_format) {
+            log::error!("Surface formats are incompatible: {:?}", supported_formats);
+        }
+
+        let vk_usage = super::resource::map_texture_usage(config.usage, crate::TexelAspects::COLOR);
+        if !capabilities.supported_usage_flags.contains(vk_usage) {
+            log::error!(
+                "Surface usages are incompatible: {:?}",
+                capabilities.supported_usage_flags
+            );
+        }
+
+        let mut full_screen_exclusive_info = vk::SurfaceFullScreenExclusiveInfoEXT {
+            full_screen_exclusive: if config.allow_exclusive_full_screen {
+                vk::FullScreenExclusiveEXT::ALLOWED
+            } else {
+                vk::FullScreenExclusiveEXT::DISALLOWED
+            },
+            ..Default::default()
+        };
+
+        let mut create_info = vk::SwapchainCreateInfoKHR {
+            surface: surface.raw,
+            min_image_count: effective_frame_count,
+            image_format: surface_format.format,
+            image_color_space: surface_format.color_space,
+            image_extent: vk::Extent2D {
+                width: config.size.width,
+                height: config.size.height,
+            },
+            image_array_layers: 1,
+            image_usage: vk_usage,
+            pre_transform: vk::SurfaceTransformFlagsKHR::IDENTITY,
+            composite_alpha,
+            present_mode,
+            old_swapchain: surface.swapchain.raw,
+            ..Default::default()
+        }
+        .queue_family_indices(&queue_families);
+
+        if self.device.full_screen_exclusive.is_some() {
+            create_info = create_info.push_next(&mut full_screen_exclusive_info);
+        } else if !config.allow_exclusive_full_screen {
+            log::info!("Unable to forbid exclusive full screen");
+        }
+        let raw_swapchain = unsafe { khr_swapchain.create_swapchain(&create_info, None).unwrap() };
+
+        unsafe {
+            self.deinit_swapchain(surface);
+        }
+
+        let images = unsafe { khr_swapchain.get_swapchain_images(raw_swapchain).unwrap() };
+        let target_size = [config.size.width as u16, config.size.height as u16];
+        let subresource_range = vk::ImageSubresourceRange {
+            aspect_mask: vk::ImageAspectFlags::COLOR,
+            base_mip_level: 0,
+            level_count: 1,
+            base_array_layer: 0,
+            layer_count: 1,
+        };
+        for image in images {
+            let view_create_info = vk::ImageViewCreateInfo {
+                image,
+                view_type: vk::ImageViewType::TYPE_2D,
+                format: surface_format.format,
+                subresource_range,
+                ..Default::default()
+            };
+            let view = unsafe {
+                self.device
+                    .core
+                    .create_image_view(&view_create_info, None)
+                    .unwrap()
+            };
+            let semaphore_create_info = vk::SemaphoreCreateInfo::default();
+            let acquire_semaphore = unsafe {
+                self.device
+                    .core
+                    .create_semaphore(&semaphore_create_info, None)
+                    .unwrap()
+            };
+            surface.frames.push(super::InternalFrame {
+                acquire_semaphore,
+                image,
+                view,
+            });
+        }
+        surface.swapchain = super::Swapchain {
+            raw: raw_swapchain,
+            format,
+            target_size,
+        };
+
+        crate::SurfaceInfo { format, alpha }
+    }
+
+    pub fn acquire_frame(&self, surface: &mut super::Surface) -> super::Frame {
+        let khr_swapchain = self.device.swapchain.as_ref().unwrap();
+        let acquire_semaphore = surface.next_semaphore;
+        match unsafe {
+            khr_swapchain.acquire_next_image(
+                surface.swapchain.raw,
+                !0,
+                acquire_semaphore,
+                vk::Fence::null(),
+            )
+        } {
+            Ok((index, _suboptimal)) => {
+                surface.next_semaphore = mem::replace(
+                    &mut surface.frames[index as usize].acquire_semaphore,
+                    acquire_semaphore,
+                );
+                super::Frame {
+                    internal: surface.frames[index as usize],
+                    swapchain: surface.swapchain,
+                    image_index: index,
+                }
+            }
+            Err(vk::Result::ERROR_OUT_OF_DATE_KHR) => {
+                log::warn!("Acquire failed because the surface is out of date");
+                super::Frame {
+                    internal: super::InternalFrame::default(),
+                    swapchain: surface.swapchain,
+                    image_index: 0,
+                }
+            }
+            Err(other) => panic!("Aquire image error {}", other),
+        }
+    }
+}
